migrate TS cutter to node
when running ffmpeg to clip the video, we need to apply the same logic we
use to transcode an ad to the stream parameters (specifically, the problem
was with AAC audio profiles). in order to avoid the need to re-implement
this logic in native C, the ts_cutter was migrated to node.
all the parts that run per-frame/per-packet were implemented in 5 native
functions exposed as a node addon, all the wrapping logic was implemented
in JS.