never output the pre/ad/post in a cyclic manner
only the filler can be cyclic.
normally, the duration of the pre/ad/post segments in the layout prevent
it from happening, however in some cases the pre-ad may start with a zero
duration frame, without this change build_layout would try to push this
frame at the end of the pre-ad, resulting in a corrupt video.